Extract ServiceRequest & ServiceRequestId from service_server.hpp
