+++ email/include/email/service_server.hpp
@@ -30,45 +30,11 @@
 #include "email/service.hpp"
 #include "email/service_handler.hpp"
 #include "email/service_info.hpp"
+#include "email/service_request.hpp"
 #include "email/visibility_control.hpp"
 
 namespace email
 {
-
-/// Service request ID, with a sequence number and a client GID.
-struct ServiceRequestId
-{
-  /// Sequence number of the request.
-  uint32_t sequence_number;
-  /// GID of the service client that made the request.
-  Gid client_gid;
-  /// Constructor.
-  ServiceRequestId(const uint32_t sequence_number_, const Gid & client_gid_)
-  : sequence_number(sequence_number_),
-    client_gid(client_gid_)
-  {}
-  /// Copy constructor.
-  ServiceRequestId(const ServiceRequestId &) = default;
-};
-
-/// Service request, with a request ID and some content.
-struct ServiceRequest
-{
-  /// Request ID.
-  ServiceRequestId id;
-  /// Content of the request.
-  std::string content;
-  /// Constructor.
-  ServiceRequest(
-    const uint32_t sequence_number_,
-    const Gid & client_gid_,
-    const std::string & content_)
-  : id(sequence_number_, client_gid_),
-    content(content_)
-  {}
-  /// Copy constructor.
-  ServiceRequest(const ServiceRequest &) = default;
-};
 
 /// Service server.
 /**
@@ -100,7 +66,7 @@ public:
    * \return the request, or `std::nullopt` if there is none
    */
   EMAIL_PUBLIC
-  std::optional<ServiceRequest>
+  std::optional<struct ServiceRequest>
   get_request();
 
   /// Get a request with info if there is one.
@@ -108,7 +74,7 @@ public:
    * \return the request with info, or `std::nullopt` if there is none
    */
   EMAIL_PUBLIC
-  std::optional<std::pair<ServiceRequest, ServiceInfo>>
+  std::optional<std::pair<struct ServiceRequest, ServiceInfo>>
   get_request_with_info();
 
   /// Send response.
@@ -118,7 +84,7 @@ public:
    */
   EMAIL_PUBLIC
   void
-  send_response(const ServiceRequestId & request_id, const std::string & response);
+  send_response(const struct ServiceRequestId & request_id, const std::string & response);
 
 private:
   EMAIL_DISABLE_COPY(ServiceServer)
